+++ modules/codec/opus_header.h
@@ -50,6 +50,6 @@ void opus_header_init(OpusHeader *);
 void opus_header_clean(OpusHeader *);
 int opus_header_parse(const unsigned char *header, int len, OpusHeader *h);
 void opus_prepare_header(unsigned channels, unsigned rate, OpusHeader *header);
-int opus_write_header(uint8_t **p_extra, int *i_extra, OpusHeader *header, const char *vendor);
+int opus_write_header(uint8_t **p_extra, size_t *i_extra, OpusHeader *header, const char *vendor);
 
 #endif
